Introduce the delivery options application
This commit introduces the "delivery options" application. This
application is designed as the reference application for the course, and
will be instrumented by the learner (as well as during the
demonstrations) to make it more "observable".

== Design Notes
=== Documentation

Given the audience of this work is new, the application tends toward
assuming very little knowledge of systems form the learner and is very
verbose.