Add the exercise for implementing the delivery options service main func
This commit introduces the exercise for the delivery service "main"
func. It is not elegant, and rather exists as a kind of a "stub" while i
figure out a more elegant way to do this.

The more I think about it, the more I am thinking it makes more sense to
write this application in its entirety rather than try and stub it out
like this.