Provide extensible mechanism for server
Currently the server is a very minimal implementation, with only a new,
start and shutdown function (which are just proxied to http.Server).
This commit extends them without breaking the API, and thus, the exiting
learning material.

It extends them by creating a property "default handlers", the server
and so on. This allows each handler to be tested, even if the server
itself cannot.

It is not anticipated the server will be especially interesting from the
perspective of the instrumentation, though it is likely there will need
to be some middleware or similar at some point.