Introduce OpenAPI specification
This commit introduces an OpenAPI specification. This is just a handy
way to remember what the implementation of this (mocked) service is
supposed to look like, as well as a way for me to visually double check
this before I commit to it.

== Design Notes
=== README

With the OpenAPI specification implemented, it is now clear in the
README what should be returned to the user.