Implement a lot of the carrier primitives
This commit implements a lot of the carrier primitives that are required
for the "happy path" of this application to work as expected.

== Design Notes
=== No tests, general design constraints

This commit is notable as we're growing the amount of functionality that
exists without tests. There's a practical and a principal reason for
this.

  1. Practical

  The practical reason is its 21:30 at night, and I sleep early. I
  go to the fitness studio in the morning, and it sucks if I sleep late.

  2. Principal

  I'm not sure how much value tests add here. Making the code testable
  means modularizing it substantially, and I'm not sure if this will
  make it harder for new go engineers to review it.

  More experience go engineers will, of course, look at my code and
  laugh. However, I'm happy to prioritize the clarity for newbies
  (and verbosity) over modularized and tested code.

  I will think on it more.

== Carrier Package

This commit provides a carrier package which provides common
infrastructure for whatever carriers will be implemented in future.

I do not expect the carriers to actually be separate. They'll all be
clones of each other, injected with an amount of entropy and
configuration to split their behavior from an observability perspective.

However, I need to isolate that mock implementation from the part that
people are expected to instrument. I will figure this out.

The money, delivery option and package structs are all implemented from
the OpenAPI spec.

== Money package

This commit implements a money package. This is overkill, but given taht
people will be critiquing this code looking for random stuff, skipping
the "floats and money bad" discussion seems simpler.

== Extended error handling in delivery options

This commit also extends the error handling in delivery options to
validate that the inputs can be converted to int64 (and does so), rather
than just validates they're present.

This is a requirement to use the "Package" struct.