Introduce the OpenTelemetry Prometheus metrics handler
The next unit is on _metrics_. In order to have a good conversation
about metrics, we need a reference implementation of them and a way to
invoke that implementation.

This commit introduces:

1. OpenTelemetry as the reference implementation for metrics, and
2. Prometheus as the exporter that will be used for this implementation.

== Design Notes
=== Pluggable Exporters

This commit provides a "pluggable" framework for exporters. There will
be multiple different points within this application that we will be
refactoring the metrics pipeline, as we get more toward something that
is "production worthy".

This might be (for example) switching out the Prometheus implementation
with an implementation like the OTel Collector implementation (or a
vendor implementation).

This commit builds the code in a way that allows this.

=== Prometheus custom handler / port

The default prometheus handler has two exporters enabled; the "go"
exporter and the "process" exporter. I think this is a little complex
for beginners, having metrics turn up when there should not be anything
available.

This commit thus opts for a simpler output, with a slightly more complex
impl.

The separate mux and port come as a consequence of this, though for
hygeine purposes thats usually a good idea anyway. If nothing else, this
port can be mTLS'd in production.

=== No tests

I am running behind on this work, and just trying to get the first draft
out before the end of october. I'll refactor it if it makes any cash,
and people actually start using it.