Factor out carriers
Currently I am working through the application, instrumenting it with
suitable timeseries that we can reason through its internal state.
However, while doing so, I was looking at the method that queries
delivery options for where to instrument the delivery logic.

The best answer was "not there". The method is getting extrodinarily
long, and is a difficult read. The instrumentation logic is verbose, and
will make that all the more challenging.

Given this, I refactored it out into its own pcakage (carriers.go),
where the instrumentation logic can live.