Introduce metrics into the carrier
Currently there are no custom metrics in the application. This is more
problematic, as the goal of this course is to teach the different places
to use the different metric types.

This commit seeks to address that by introducing metric based visibility
into the "carriers" process. This requires a substantial refactor of the
carriers process to ensure that it is injectable. The specific design
notes are below:

== Design Notes
=== Splitting carrier from carriers

Currently both the "carrier" interfaces and "carriers" options are
within the same file. This gets a little tricky, as the "carriers" logic
is now gettin guqite complex.

This commit splits the "carrier" specific logic — the carrier interface,
and corresponding structs — into its own file with the clear purpose.
This allows this material to be colocated, as well as to still have the
clear separation between the "carrier" and the "carriers", but also to
enable focus within the "carriers" file on just the logic for multiple
carriers.

=== Varidic Constructor

The required logic to correctly create the Carriers method is getting
complex. Given this, a constructor has been added to take a lot of this
complexity away behind a simpler API.

This commit implements the variadic constructor pattern popularized by
Dave Cheany and Uber¹. This allows an extensible set of options, without
needing to modify the APIs down the road.

This commit also provides a "Defaults" option which is expected to be
used and extended in most cases, but which is not a hard requirement.
This makes it hard to use wire based dependency injection; I haven't
solved that just yet.
—
1. https://dave.cheney.net/2014/10/17/functional-options-for-friendly-apis

=== Opts and Metrics sub-structs

There are anonymous structs in the go for both the options (used only at
boot time) as well as the metrics (used both at boot time and run time,
but not part of the "core" carriers logic).

The goal of using anonymous structs here is to communicate that while
they're part of the "carriers" method, they're not really part of the
"core" logic of the method.

=== Server fixes

While in previous commits I've been continuing to refactor this work to
make it easier to encapsulate more logic (including instrumentation),
the actual code had a bug in it — dual defined servers — which has been
cleaned up in this commit.