Introduce the examples common setup
Distributed tracing is quite a bit more complicated than logging in that
it requires at least some bootstrap. There's a lot of logic that can be
executed in-application, such as sampling and processing and so on, so
this bootstrap can be pretty complicated.

This commit does the setup for a local instance of tracing such that we
can proceed with the course in a way users can be interactive with, but
that is simplified.

== Design Notes
=== General Setup

This is pretty complicated. OTel requires a bunch of complexity to
"break out" of production readiness into a local develompent mode (e.g.
setting up gRPC without TLS)

This is frustrating for learners, but something we can later abstract
behind a package.

=== Jaeger

The courses goes through it, but essentially the code expects a local
jaeger instance with an exposed gRPC connection.

=== Returning tracer provider

As far as I can reason, the best "pattern" for tracing is to have a
given package:

1. Init to a global provider (either via a package variable or, more
   likely, a struct property)
2. Allow passing in a specific provider (e.g. a testable one)

The simplest way in the case of both of these is to have that bootstrap
deal with the provider, rather than expect a created tracer, as the
library can then specify its own name (and so on).

Given this, this is what I'll teach.