Introduce lecture issue template
The next stream of work will be the "Metrics" work. The broad outlines
of this work are not yet complete, but right now my goal is to create a
series of "primitives" to "stack" into a unit. Things like lectures,
quizzes, exercises, cookbooks and so on.

Then, can just "assemble" the unit, and ship.