Split the user-facing from internal issues with emoji
This project uses GitHub as the issue managment tooling. To keep
progress transparent for enthousiasts, its all public. However, because
of the way I'm constructing the courses, this means this is all getting
a bit confusing.

This commit attempts to provide some split by taking advantage of
GitHubs alphabetic sort of issues. Prefixing the issues with an emoji
means that we can "group" the "user facing" and "internal" issues.
Ideally the user facing ones would come first, but lets see how it
appears.