Revert "Use letter prefixes"
GitHub uses the file name as the ordering mechanism, rather than the
name of the content. The fact it "seemed" to be alphabetical is a
consequence if me not looking properly and chance¹

This commit reverts the ascii changes, and keeps the emoji ones for
visual separation. It also sorts the files by prefixing them with
numbers.

This reverts commit c1b32caf52073df3598db660e173936cab3f0435.
—
1. https://github.com/orgs/community/discussions/21601