Prefix the shorts with "Lets talk about ..."
Recently, I was down the "YouTube" shorts rabbit hole, and one of the
more compelling hooks was prefixing the videos with "I work on a Cruise
ship and ..."

This commit attempts to replicate this with "Lets talk about ...", along
with the segments "Observability" or "The observability course". The
goal being to create a similar audio brand, as well as give people an
idea as to what follows for the next 55 seconds or so. Lastly, to allow
the title to be used.

There are checkboxes that link the prefix with hashtags, so delimit the
various conditions.