Introduce story primitive
A story is just an example from my life historically about where these
technologies, tools or processes have been used. It's designed foremost
as a mental break and a kind of entertainment with loose moral that
applies to the course.

Secondarily, it serves to reinforce or outline some specific point.

This commit introuces a template for a course section like this.