Add practicals (originally: cookbooks)
There are a set of things that are kind of, specific sets of
instructions for people to do in order to drive their organization to a
desired outcome. These are "Reliability Enhancement Procedures" (see
Swisscom), "Cookbooks" or "Practicals" depending on the context. Here,
the name "practical" is chosen as "Prac" reminds me of school.

This commit implements an example practical document.