Introduce an initial pass a the metrics setup file
Like the tracing example, the metrics implementation requires quite a
bit of setup that's easier to cover once, and then omit from specific
files. This commit implements that setup.

== Design Notes
=== Passing in the service name

Unlike the tracing example, here, we're passing in the service name to
the NewMetricProvider function. This allows us to more easily delimit
examples than the previous tracing design, which should probably be
refactored to use a similar approach.

Long term, I want to pull these out of the course and write a decent
bootstrap library that just takes care of this. It's not really relevant
for learners, and its always good to remove content from the course.

=== Passing back the reader

As mentioned in the comments, the reader allows us to flush the metrics
manually.