Introduce the Go GC metric count
This commit introduces a new metric, which is a count of the number of
times Go executes its garbage collection. This is designed just to
provide some examples of how to query runtime metrics, and to expand the
complexity of the topic so we can introduce some different primitives.

== Design Notes
=== Increasing collection frequency

This commit increases the collection frequency from 60s → 15s so that
the example can run for 60s and produce some sort of meaningful data.

=== Naming examples

The examples are built on top of each other, which means a lot of copy /
paste code. To make this easier to read, we're just substituting the
names for variable.

=== GC as the metric

Originally, the goal was to have runtime.uptime as the metric, but
there's some discussion this should be an async gauge rather than a
counter. So as not to confuse the issue, I Just picked something that is
more clearly represented as a counter.