Introduce an example reading statistics from PSI
This commit introduces a (limited) example reading the current
statistics from /proc/PSI. This allows an example of exporting machine
state via OpenTelemetry, and then querying it back.

== Design Notes
=== General Design

The general design is a first pass. In particular, I really do not like
the scope leaking of the closure to capture the state; thus the passing
in of the PSI<blah> variables.

Additionally, the file handles should probably just remain open and
periodically be completely re-read. They do not change handle between
uses or anything else similar.