Refactor the implementation of PSI
This commit tidies the implementation of PSI a little. Specifically, it
removes some of the maps that were present, and schedules an execution
for each "form" of pressure differently.

== Design Notes
=== Noop Error

I'm not super happy with the function that creates the "observer"
functions still, but it is sufficiently clear that it will have to do.