Pass a single context through multiple methods
While the specifics of the usage of Context() is not clear, debugging
through the code indicates that an error can get attached to the
context, which is subsequently checked. With multiple calls, this is
likely going to avoid situation where the metrics are recorded when the
previous failed.

I need to read further on this.